{{ cookiecutter.name }}/backend/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.name }}"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
psycopg2 = "^2.9.5"
django-redis-cache = "^3.0.1"
celery = "^5.2.7"
django-debug-toolbar = "^3.7.0"
gunicorn = "^20.1.0"
factory-boy = "^3.2.1"
django-filter = "^22.1"
drf-spectacular = "^0.24.2"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
